Add GSEA analysis; remove benchmarking of goSeq analysis
Adding GSEA analysis (by fgsea package) which implements a FCS
(Functional Class Scoring) method. Unlike goSeq, it does not
require an arbitrary threshold to define DE genes. Instead it
uses the rankings from a DE analysis, and checks for enrichment
of these rankings for all genes. Here the gene set size appears
to be both informative and independent under the null, since
we no longer have to perform selection/filtering of the p-values.

Also removing the benchmarking from the goSeq analyses. This is
because we see that (and logically intuit that) the assumption
of indpendent p-value under the null by different covariate
values is violated. Added detailed explanation as to why.